requirements.txt
discord.py==1.7.3
asyncpraw==7.5.0
asyncprawcore==2.3.0
python-dotenv==0.19.2
table2ascii==0.1.4
pytz==2021.3
dateparser==1.1.0